Added camera usage, multiple bugfixes to shaders and framebuffers

+++ GUI/Camera.cs
@@ -72,7 +72,7 @@ namespace GUI
         {
             get
             {
-                return screenWidth / screenHeight;
+                return screenWidth / (double)screenHeight;
             }
         }
         /// <summary>
